+++ src/store/games/gamesStore.ts
@@ -1,11 +1,11 @@
-import { CONFIG_DEFAULT_COLLECTIONS_MENU_SLUGS } from "@theme/configs/categoryesGames";
-import { SlugCategoriesGames } from "@theme/configs/categoryesGames";
+import { CONFIG_DEFAULT_COLLECTIONS_MENU_SLUGS, SlugCategoriesGames } from "@theme/configs/categoryesGames";
 import { defineStore, type Pinia, storeToRefs } from "pinia";
 import { computed, ref } from "vue";
 
 import { log } from "../../controllers/Logger";
 import { currencyView } from "../../helpers/currencyHelper";
 import { processGame } from "../../helpers/gameHelpers";
+import { isExistData } from "../../helpers/isExistData";
 import type { IGame, IGamesProvider } from "../../models/game";
 import { IEnabledGames } from "../../models/game";
 import {
@@ -159,6 +159,9 @@ export const useGamesCommon = defineStore("gamesCommon", () => {
 
     async function loadGamesCategories(): Promise<void> {
         try {
+            if (isExistData(gamesCategories.value)) {
+                return;
+            }
             const data = await loadGamesCategoriesReq();
             gamesCategories.value = data.map((category) => {
                 return {