+++ src/store/games/helpers/games.ts
@@ -4,19 +4,37 @@ import { storeToRefs } from "pinia";
 import { IGameItem } from "../../../helpers/gameHelpers";
 import type { ICollectionItem, IGame, IGamesProvider } from "../../../models/game";
 import { GameDisableGeoStatus, IEnabledGames } from "../../../models/game";
+import { useRootStore } from "../../../store/root";
 import { useCommon } from "../../common";
 import { useContextStore } from "../../context";
 import { useGamesProviders } from "../gamesProviders";
 import { useGamesCommon } from "../gamesStore";
 
-// tslint:disable-next-line:max-line-length
-function findGameBySeoTittleAndProducerWithDuplicate(gamesCollection: IGame[], { producer, seoTitle }): IGame | undefined {
-    return gamesCollection.find(({ seo_title: seoTitleItem, provider: providerItem }) => {
-        return producer === providerItem && seoTitleItem === seoTitle;
-    });
+
+type IFindGameParams = {
+    producer: string;
+    seoTitle: string;
+};
+
+function findGameBySeoTittleAndProducerWithDuplicate(
+    gamesCollection: IGame[],
+    { producer, seoTitle }: IFindGameParams,
+): IGame | undefined {
+    const rootStore = useRootStore();
+    return gamesCollection
+        .filter((game: IGame) => {
+            return game.devices.length > 1 ||
+                game.devices.includes(rootStore.isMobile ? "mobile" : "desktop");
+        })
+        .find(({ seo_title: seoTitleItem, provider: providerItem }) => {
+            return producer === providerItem && seoTitleItem === seoTitle;
+        });
 }
 
-export function findGameBySeoTittleAndProducer(gamesCollection: IGame[], { producer, seoTitle }): IGame | undefined {
+export function findGameBySeoTittleAndProducer(
+    gamesCollection: IGame[],
+    { producer, seoTitle }: IFindGameParams,
+): IGame | undefined {
     return findGameBySeoTittleAndProducerWithDuplicate(gamesCollection, { producer, seoTitle });
 }