+++ src/store/lotteries.ts
@@ -52,6 +52,7 @@ export const useLotteriesStore = defineStore("lotteriesStore", () => {
     }
 
     return {
+        lotteriesList,
         lotteriesStatuses,
         setLotteriesStatuses,
         currentLottery,