+++ CHANGELOG.md
@@ -1,3 +1,151 @@
+## [1.49.0](https://github.com/upstars-global/unity-core-modules/compare/v1.48.2...v1.49.0) (2025-11-17)
+
+### 🚀 Features
+
+* **UN-1775:** run terms popup in a main product ([#244](https://github.com/upstars-global/unity-core-modules/issues/244))
+ ([69acd24](https://github.com/upstars-global/unity-core-modules/commit/69acd24fa3d0e471cc0a8f29ff1a17af001b3629))
+
+
+
+    * feat(UN-1775): run terms popup in a main product
+
+    * fix: Pass the accept terms handler as a prop to the accept terms modal.
+
+## [1.48.2](https://github.com/upstars-global/unity-core-modules/compare/v1.48.1...v1.48.2) (2025-11-14)
+
+### 🐛 Bug Fixes
+
+* minor changeUserToGroup
+ ([ade8451](https://github.com/upstars-global/unity-core-modules/commit/ade8451973c0407facd72d6ad2eefe20faf33eb4))
+
+
+
+    UN-1902: fix 422 по api/groups
+
+## [1.48.1](https://github.com/upstars-global/unity-core-modules/compare/v1.48.0...v1.48.1) (2025-11-14)
+
+### 🐛 Bug Fixes
+
+* 422 api/groups and minor refactor
+ ([24d92dd](https://github.com/upstars-global/unity-core-modules/commit/24d92dd6caa5204be7647b16484b734e68c3c8e5))
+
+
+
+    UN-1902: 422 по api/groups
+
+## [1.48.0](https://github.com/upstars-global/unity-core-modules/compare/v1.47.0...v1.48.0) (2025-11-12)
+
+### 🚀 Features
+
+* add log for /api/player/groups
+ ([57cca86](https://github.com/upstars-global/unity-core-modules/commit/57cca8621844029daf6798628d1600667aebc0e3))
+
+, closes [#256](https://github.com/upstars-global/unity-core-modules/issues/)
+
+## [1.47.0](https://github.com/upstars-global/unity-core-modules/compare/v1.46.1...v1.47.0) (2025-11-12)
+
+### 🚀 Features
+
+* Un 1776 ([#249](https://github.com/upstars-global/unity-core-modules/issues/249))
+ ([6bfc95b](https://github.com/upstars-global/unity-core-modules/commit/6bfc95b28c7aa4ac41bcc503de0e722bff18115c))
+
+
+
+    * UN-1776 mart for king
+
+    * UN-1776 mart for king
+
+    * UN-1874 UN-1871 add sort for comp point rates
+
+    * UN-1874 UN-1871 add sort for comp point rates
+
+    * UN-1776 disable rates sort
+
+## [1.46.1](https://github.com/upstars-global/unity-core-modules/compare/v1.46.0...v1.46.1) (2025-11-10)
+
+### 🐛 Bug Fixes
+
+* add filter for getDepositGiftGroupID ([#255](https://github.com/upstars-global/unity-core-modules/issues/255))
+ ([9a068df](https://github.com/upstars-global/unity-core-modules/commit/9a068df9afaaffb8a286db0f33edc21c63ebb667))
+
+## [1.46.0](https://github.com/upstars-global/unity-core-modules/compare/v1.45.0...v1.46.0) (2025-11-10)
+
+### 🚀 Features
+
+* add choosing bonus cashbox
+ ([6a5063e](https://github.com/upstars-global/unity-core-modules/commit/6a5063e1aa96746231a34810edd0c96163089f55))
+
+
+
+    https://upstars.atlassian.net/browse/UN-1733
+
+## [1.45.0](https://github.com/upstars-global/unity-core-modules/compare/v1.44.2...v1.45.0) (2025-11-07)
+
+### 🚀 Features
+
+* **UN-1804:** Replace payment endpoint with pagination ([#250](https://github.com/upstars-global/unity-core-modules/issues/250))
+ ([fdc5530](https://github.com/upstars-global/unity-core-modules/commit/fdc5530a3b862e2d5b25fff3315133b1f44fc263))
+
+
+
+    * UN-1804 WIP
+
+    * UN-1804 WIP
+
+    * UN-1804
+
+    * UN-1804
+
+    * UN-1804
+
+    * UN-1804 correct pagination size
+
+    * UN-1804 default 50 size
+
+    * UN-1804 logs
+
+    * UN-1804 change to 20
+
+    * UN-1804
+
+    * UN-1804
+
+## [1.44.2](https://github.com/upstars-global/unity-core-modules/compare/v1.44.1...v1.44.2) (2025-11-06)
+
+### 🐛 Bug Fixes
+
+* UN-1874 fix disabled provider config ([#254](https://github.com/upstars-global/unity-core-modules/issues/254))
+ ([1f52704](https://github.com/upstars-global/unity-core-modules/commit/1f527041a35e6a4912d4f89b27e205e3c657e48c))
+
+## [1.44.1](https://github.com/upstars-global/unity-core-modules/compare/v1.44.0...v1.44.1) (2025-11-06)
+
+### ⏪ Reverts
+
+* Revert "[#242](https://github.com/upstars-global/unity-core-modules/issues/242)" ([#252](https://github.com/upstars-global/unity-core-modules/issues/252))
+ ([eb29c91](https://github.com/upstars-global/unity-core-modules/commit/eb29c9190502ce0152c01f368c6334f531bb110f))
+
+
+
+### 🐛 Bug Fixes
+
+* filter games devices by user device ([#242](https://github.com/upstars-global/unity-core-modules/issues/242))
+ ([5d39f00](https://github.com/upstars-global/unity-core-modules/commit/5d39f00cddedaf8ea5bb9d21e0b65c7985597fb8))
+
+
+
+    * fix: filter games by device
+
+    * fix: game orientation detection
+* **UN-1797:** filter games devices by user device ([#253](https://github.com/upstars-global/unity-core-modules/issues/253))
+ ([2aaf465](https://github.com/upstars-global/unity-core-modules/commit/2aaf465930abb837d3e77ab9b1ece7bfb238a76d))
+
+## [1.44.0](https://github.com/upstars-global/unity-core-modules/compare/v1.43.1...v1.44.0) (2025-11-05)
+
+### 🚀 Features
+
+* Un 113 ([#248](https://github.com/upstars-global/unity-core-modules/issues/248))
+ ([86fcfae](https://github.com/upstars-global/unity-core-modules/commit/86fcfae6bc41701b30cd237c91ad92ed1d8b1fd9))
+
 ## [1.43.1](https://github.com/upstars-global/unity-core-modules/compare/v1.43.0...v1.43.1) (2025-10-30)
 
 ### ⏪ Reverts